Added visual comparison test and example screens/report
Added the Visual_JobPage.js file which performs the visual comparison tests on the job page.

Also kept in the screenshots in the VisualComparisonScreenshots folder, and the data / report in the "generated" folder. Kept these in as an example of how it all looks after running all 3 of the commands blink-diff requires.