Final Tidy Up
Realised that the job sector pages had issues. Sometimes the <search button was in a different location so introduced the inbuilt testcafe function to store navigation history and go back one page.

Also realised that the apply button could sometimes have a different selector, so swapped to seaching the container rather than the selector.

Ran test suite in full and generated final example report which replaces the previous one